refactor(common router): move router to common
